Added BanDetector to Server
- Uses a more OOP style (will apply that style to all others)
